refact: improve performance of saving pool state
