+++ client/src/pages/SwapPage/index.tsx
@@ -1,4 +1,3 @@
-import { useAtom } from 'jotai';
 import { useMutation, useQuery } from "react-query";
 import { useNavigate } from "react-router-dom";
 import { MdSwapCalls } from "react-icons/md";
@@ -14,13 +13,12 @@ import { sleep } from "src/lib/utils";
 
 import { queryPreviewAmount, swapTokens } from "./queries";
 import { SwapComponent } from "./SwapComponent";
-import { swapAtom } from "./jotai";
+import { SwapState } from "./types";
 
 export default function SwapPage() {
   const contract = useContract()!;
   const [previewAmount, setPreviewAmount] = useState<bigint | null>(null);
-  const [swapState, setSwapState] = useAtom(swapAtom);
-
+  const [swapState, setSwapState] = useState<SwapState | null>(null);
   const [hasLiquidity, setHasLiquidity] = useState(true);
   const debouncedState = useDebounce(swapState);
   const navigate = useNavigate();
@@ -74,10 +72,7 @@ export default function SwapPage() {
         <MdSwapCalls className="text-primary-500" />
         Swap
       </PageContent.Title>
-      <SwapComponent
-        amount={debouncedState?.amount}
-        previewAmount={previewAmount}
-        onChange={setSwapState} />
+      <SwapComponent previewAmount={previewAmount} onChange={setSwapState} />
       <Button
         isFull
         size="lg"
