+++ client/src/pages/SwapPage/types.ts
@@ -1,0 +1,11 @@
+export enum ActiveInput {
+  'from' = 'from',
+  'to' = 'to',
+}
+
+export type SwapState = {
+  from: string;
+  to: string;
+  direction: ActiveInput;
+  amount: bigint | null;
+};
